images_processing_service/pyproject.toml
[tool.poetry]
name = "images-processing-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
faststream = "^0.5.39"
dishka = "^1.5.3"
uvicorn = "^0.34.2"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
psycopg = "^3.2.6"
environs = "^14.1.1"
aio-pika = "^9.5.5"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
